Add ISPC backend

+++ vlm/include/vlm.hpp
@@ -19,19 +19,19 @@ class Solver {
 class NonLinearVLM final: public Solver {
     public:
     const f64 tol;
-    const u32 max_iter;
+    const u64 max_iter;
     std::vector<f32> strip_alphas;
 
     NonLinearVLM(
         const tiny::Config& cfg,
         const f64 tol_= 1e-4f,
-        const u32 max_iter_ = 100
+        const u64 max_iter_ = 100
     ):
         Solver(cfg),
         tol(tol_),
         max_iter(max_iter_)
     {}
-    NonLinearVLM(const f64 tol_= 1e-4f, const u32 max_iter_ = 100) : tol(tol_), max_iter(max_iter_) {}
+    NonLinearVLM(const f64 tol_= 1e-4f, const u64 max_iter_ = 100) : tol(tol_), max_iter(max_iter_) {}
     ~NonLinearVLM() = default;
     AeroCoefficients solve(const FlowData& flow, const Database& db);
 };
