+++ vlm/include/vlm_backend.hpp
@@ -16,11 +16,11 @@ class Backend {
         virtual void compute_rhs(const FlowData& flow, const std::vector<f32>& section_alphas) = 0; 
         virtual void lu_factor() = 0;
         virtual void lu_solve() = 0;
-        virtual f32 compute_coefficient_cl(const FlowData& flow, const f32 area, const u32 j, const u32 n) = 0;
+        virtual f32 compute_coefficient_cl(const FlowData& flow, const f32 area, const u64 j, const u64 n) = 0;
         f32 compute_coefficient_cl(const FlowData& flow);
-        virtual linalg::alias::float3 compute_coefficient_cm(const FlowData& flow, const f32 area, const f32 chord, const u32 j, const u32 n) = 0;
+        virtual linalg::alias::float3 compute_coefficient_cm(const FlowData& flow, const f32 area, const f32 chord, const u64 j, const u64 n) = 0;
         linalg::alias::float3 compute_coefficient_cm(const FlowData& flow);
-        virtual f32 compute_coefficient_cd(const FlowData& flow, const f32 area, const u32 j, const u32 n) = 0;
+        virtual f32 compute_coefficient_cd(const FlowData& flow, const f32 area, const u64 j, const u64 n) = 0;
         f32 compute_coefficient_cd(const FlowData& flow);
         virtual void compute_delta_gamma() = 0;
         virtual ~Backend() = default;
